refactor parse related files and put them in parse directry (fix some error messages)
