--- src/color.h
@@ -1,20 +1,0 @@
-/*#ifndef COLOR_H
-# define COLOR_H
-
-
-struct s_color {
-	FLOAT	r;
-	FLOAT	g;
-	FLOAT	b;
-};
-
-t_color	color_black(void);
-t_color	color_white(void);
-t_color	color_new(FLOAT x, FLOAT y, FLOAT z);
-t_color	color_add(t_color a, t_color b);
-t_color	color_sub(t_color a, t_color b);
-t_color	color_mul(t_color a, t_color b);
-int	get_mlx_color(t_color c);
-
-
-#endif*/
