+++ src/img.c
@@ -6,7 +6,7 @@
 /*   By: katakagi <katakagi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/01 12:08:02 by susami            #+#    #+#             */
-/*   Updated: 2023/02/09 21:22:19 by katakagi         ###   ########.fr       */
+/*   Updated: 2023/02/10 03:00:04 by katakagi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ void	*init_img(void	*mlx_ptr, int width, int height)
 
 	img = ft_calloc(1, sizeof(*img));
 	if (img == NULL)
-		fatal_error("init img", NULL);
+		fatal_error("init img", "calloc error");
 	img->ptr = mlx_new_image(mlx_ptr, width, height);
 	if (img->ptr == NULL)
 		exit(1);
