+++ src/parse/element.c
@@ -6,13 +6,12 @@
 /*   By: katakagi <katakagi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/08 23:57:58 by katakagi          #+#    #+#             */
-/*   Updated: 2023/02/09 23:02:49 by katakagi         ###   ########.fr       */
+/*   Updated: 2023/02/10 02:50:41 by katakagi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minirt.h"
-#include "tokenize.h"
 #include "parse.h"
+#include "minirt.h"
 
 void	element_add(t_element *head, t_element *elem)
 {
@@ -45,38 +44,38 @@ bool	is_acl(const t_token *tok)
 	return (false);
 }
 
-// Validate orientation vector: must be normalized
-// Validate FOV: must be in range [0, 180]
-// Validate color: must be in range [0, 255]
-// Validate light ratio: must be in range [0, 255]
-void	validate_element(t_element *elem)
-{
-	const t_element_type	type = elem->element_type;
+// // Validate orientation vector: must be normalized
+// // Validate FOV: must be in range [0, 180]
+// // Validate color: must be in range [0, 255]
+// // Validate light ratio: must be in range [0, 255]
+// void	validate_element(t_element *elem)
+// {
+// 	const t_element_type	type = elem->element_type;
 
-	if (E_AMBIENT_LIGHTNING <= type && type <= E_CYLINDER && type != E_CAMERA)
-		expect_color(&elem->color);
-	if (type == E_AMBIENT_LIGHTNING || type == E_LIGHT || type == E_CAMERA)
-		expect_once(elem->element_type);
-	if (elem->element_type == E_AMBIENT_LIGHTNING)
-		expect_ratio(elem->ambient_lightning_ratio);
-	if (elem->element_type == E_CAMERA)
-	{
-		expect_normalized(&elem->orientation);
-		expect_fov(elem->hfov);
-	}
-	if (elem->element_type == E_LIGHT)
-		expect_ratio(elem->light_brightness_ratio);
-	if (elem->element_type == E_SPHERE)
-		expect_non_negative(elem->sp_diameter);
-	if (elem->element_type == E_PLANE)
-		expect_normalized(&elem->pl_normal);
-	if (elem->element_type == E_CYLINDER)
-	{
-		expect_normalized(&elem->cy_orientation);
-		expect_non_negative(elem->cy_diameter);
-		expect_non_negative(elem->cy_height);
-	}
-}
+// 	if (E_AMBIENT_LIGHTNING <= type && type <= E_CYLINDER && type != E_CAMERA)
+// 		expect_color(&elem->color);
+// 	if (type == E_AMBIENT_LIGHTNING || type == E_LIGHT || type == E_CAMERA)
+// 		expect_once(elem->element_type);
+// 	if (elem->element_type == E_AMBIENT_LIGHTNING)
+// 		expect_ratio(elem->ambient_lightning_ratio);
+// 	if (elem->element_type == E_CAMERA)
+// 	{
+// 		expect_normalized(&elem->orientation);
+// 		expect_fov(elem->hfov);
+// 	}
+// 	if (elem->element_type == E_LIGHT)
+// 		expect_ratio(elem->light_brightness_ratio);
+// 	if (elem->element_type == E_SPHERE)
+// 		expect_non_negative(elem->sp_diameter);
+// 	if (elem->element_type == E_PLANE)
+// 		expect_normalized(&elem->pl_normal);
+// 	if (elem->element_type == E_CYLINDER)
+// 	{
+// 		expect_normalized(&elem->cy_orientation);
+// 		expect_non_negative(elem->cy_diameter);
+// 		expect_non_negative(elem->cy_height);
+// 	}
+// }
 
 t_element	*elementize(const t_token *tok)
 {
@@ -98,7 +97,7 @@ t_element	*elementize(const t_token *tok)
 		else if (tok->type == TK_ID && tok->id == E_CYLINDER)
 			cur->next = parse_cylinder((const t_token **)&tok, tok);
 		else
-			fatal_error("parse", "Unexpected token");
+			fatal_error("elementize", "Unexpected token");
 		validate_element(cur->next);
 		cur = cur->next;
 	}
