+++ src/parse/element_alloc_ACL.c
@@ -6,13 +6,12 @@
 /*   By: katakagi <katakagi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/08 22:52:28 by katakagi          #+#    #+#             */
-/*   Updated: 2023/02/09 21:13:33 by katakagi         ###   ########.fr       */
+/*   Updated: 2023/02/10 02:48:35 by katakagi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-//#include "parse.h"
 #include "minirt.h"
-#include "parse.h"
+#include "element.h"
 
 t_element	*light_element_alloc(t_vec p, FLOAT ratio, t_color color)
 {
@@ -20,7 +19,7 @@ t_element	*light_element_alloc(t_vec p, FLOAT ratio, t_color color)
 
 	self = ft_calloc(1, sizeof(*self));
 	if (self == NULL)
-		fatal_error("parse", NULL);
+		fatal_error("light_element_alloc", "calloc error");
 	self->element_type = E_LIGHT;
 	self->light_point = p;
 	self->light_brightness_ratio = ratio;
@@ -34,7 +33,7 @@ t_element	*camera_element_alloc(t_vec p, t_vec dir, FLOAT hfov)
 
 	self = ft_calloc(1, sizeof(*self));
 	if (self == NULL)
-		fatal_error("parse", NULL);
+		fatal_error("camera_element_alloc", "calloc error");
 	self->element_type = E_CAMERA;
 	self->view_point = p;
 	self->orientation = dir;
@@ -48,7 +47,7 @@ t_element	*ambient_element_alloc(FLOAT ratio, t_color color)
 
 	self = ft_calloc(1, sizeof(*self));
 	if (self == NULL)
-		fatal_error("parse", NULL);
+		fatal_error("ambient_element_alloc", "calloc error");
 	self->element_type = E_AMBIENT_LIGHTNING;
 	self->ambient_lightning_ratio = ratio;
 	self->color = color;
