+++ src/parse/parse.c
@@ -6,47 +6,14 @@
 /*   By: katakagi <katakagi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/01 10:31:52 by susami            #+#    #+#             */
-/*   Updated: 2023/02/09 21:29:54 by katakagi         ###   ########.fr       */
+/*   Updated: 2023/02/10 02:52:21 by katakagi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "minirt.h"
+#include "parse.h"
 #include <fcntl.h>
 #include <errno.h>
-#include <string.h>
-#include <unistd.h>
-#include <math.h>
-#include <float.h> // FLT_EPSILON
-#include "minirt.h"
-#include "tokenize.h"
-#include "parse.h"
-
-void	expect_once(t_element_type type)
-{
-	static bool	has_seen_ambient = false;
-	static bool	has_seen_light = false;
-	static bool	has_seen_camera = false;
-
-	if (type == E_AMBIENT_LIGHTNING)
-	{
-		if (has_seen_ambient)
-			fatal_error("expect_once", "A can only be declared once.");
-		has_seen_ambient = true;
-	}
-	else if (type == E_LIGHT)
-	{
-		if (has_seen_light)
-			fatal_error("expect_once", "L can only be declared once.");
-		has_seen_light = true;
-	}
-	else if (type == E_CAMERA)
-	{
-		if (has_seen_camera)
-			fatal_error("expect_once", "C can only be declared once.");
-		has_seen_camera = true;
-	}
-	else
-		fatal_error("expect_once", "Unexpected type");
-}
 
 static char	*read_all_content(const char *filename)
 {
@@ -58,7 +25,7 @@ static char	*read_all_content(const char *filename)
 
 	fd = open(filename, O_RDONLY);
 	if (fd < 0)
-		fatal_error("open file", NULL);
+		fatal_error("read_all_content", "open file error");
 	ret = ft_calloc(1, 1);
 	while (1)
 	{
@@ -68,7 +35,7 @@ static char	*read_all_content(const char *filename)
 		buf[rc] = '\0';
 		temp = ft_strjoin(ret, buf);
 		if (temp == NULL)
-			fatal_error("strjoin", "malloc error");
+			fatal_error("read_all_content", "strjoin error");
 		free(ret);
 		ret = temp;
 	}
