+++ src/parse/parse_ACL.c
@@ -6,11 +6,10 @@
 /*   By: katakagi <katakagi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/09 00:35:29 by katakagi          #+#    #+#             */
-/*   Updated: 2023/02/09 23:02:57 by katakagi         ###   ########.fr       */
+/*   Updated: 2023/02/10 02:50:52 by katakagi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minirt.h"
 #include "parse.h"
 
 t_element	*parse_ambient_light(const t_token **rest, const t_token *tok)
@@ -68,5 +67,6 @@ t_element	*parse_acl(const t_token **rest, const t_token *tok)
 		return (parse_camera(rest, tok));
 	else if (id == E_LIGHT)
 		return (parse_light(rest, tok));
-	return (NULL);
+	else
+		return (0);
 }
