+++ src/parse/parse_other.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   element_new_other.c                                :+:      :+:    :+:   */
+/*   parse_other.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: katakagi <katakagi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/09 00:35:29 by katakagi          #+#    #+#             */
-/*   Updated: 2023/02/09 20:20:03 by katakagi         ###   ########.fr       */
+/*   Updated: 2023/02/10 02:34:54 by katakagi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
