+++ src/parse/translate2.c
@@ -6,11 +6,12 @@
 /*   By: katakagi <katakagi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/02 11:37:14 by katakagi          #+#    #+#             */
-/*   Updated: 2023/02/09 21:33:04 by katakagi         ###   ########.fr       */
+/*   Updated: 2023/02/10 02:43:06 by katakagi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
+
 
 void	translate_ambient(t_scene *scene, t_element *elem)
 {
